requirements.txt
fastapi[standard]
SQLAlchemy
passlib
bcrypt==4.0.1
python-multipart
python-jose[cryptography]
